Move Api calls to own class
