+++ Client/Components/DashBoard.razor.css
@@ -3,16 +3,16 @@
     width: 100%;
     background-color: #eeecec;
     padding: 25px;
-    display: flex;
     border-radius: 25px;
 }
 
 .data-card {
-    width: 100%;
+    width: 80%;
     height: 200px;
     border-radius: 25px;
     background-color: #f9f9f9;
     box-shadow: 0px 2px 5px 1px rgba(0,0,0,0.44);
     -webkit-box-shadow: 0px 2px 5px 1px rgba(0,0,0,0.44);
     -moz-box-shadow: 0px 2px 5px 1px rgba(0,0,0,0.44);
+    margin: 0 auto 15px auto;
 }
